Changed to use lmp_serial or lmp_mpi based on environment variables.
Changed the LAMMPS run to use lmp_serial or lmp_mpi and mpirun, based
on the value of SEAMM_LAMMPS_MPI_NP or SEAMM_MPI_NP, in that order. If
the value exists and is larger than 1, then use the parallel
version. Otherwise the serial version is used.

This supports run LAMMPS installed from conda-forge along with
openmpi.